refactor: todaySchedule, monthSchedules 훅 위치 변경
