+++ src/hooks/home/useMonthSchedule.ts
@@ -8,6 +8,7 @@ import { useAppDispatch } from "@redux/hooks.ts";
 import moment from "moment/moment";
 import { useUser } from "@app/tanstack-query/useUser.ts";
 import { useMonthSchedules } from "@app/tanstack-query/home/useMonthSchedules.ts";
+import { Schedule } from "@app/types/schedule.ts";
 
 const useMonthSchedule = () => {
   const dispatch = useAppDispatch();
@@ -26,6 +27,35 @@ const useMonthSchedule = () => {
     calendar_date: date,
   });
 
+  const schedules = monthData?.data;
+
+  const todaySchedules =
+    schedules?.filter(
+      (schedule) =>
+        moment(date).isSameOrAfter(schedule.start_date) &&
+        moment(date).isSameOrBefore(schedule.end_date)
+    ) ?? [];
+
+  const lastDay = moment(date).endOf("month").date();
+  const arr = Array.from({ length: lastDay }, (_, i) => i + 1);
+  const init: { [key: string]: Schedule[] } = {};
+  const monthSchedules = arr.reduce((prev, current) => {
+    const date = moment(`${month}-${current}`, "YYYY-MM-D");
+    const today = schedules?.filter(
+      (schedule: Schedule) =>
+        date.isSameOrAfter(schedule.start_date) &&
+        date.isSameOrBefore(schedule.end_date)
+    );
+    if (today && today.length !== 0) {
+      return {
+        ...prev,
+        [date.format("YYYY-MM-DD")]: today,
+      };
+    } else {
+      return prev;
+    }
+  }, init);
+
   const changeDate = (value: moment.Moment | null) =>
     value && dispatch(setSelectedDate(value.format("YYYY-MM-DD")));
 
@@ -34,6 +64,8 @@ const useMonthSchedule = () => {
     isPending,
     isError,
     date,
+    todaySchedules,
+    monthSchedules,
     changeDate,
   };
 };
