+++ src/hooks/useSchedule.ts
@@ -4,9 +4,8 @@ import {
   selectMonth,
   setDrawerScheduleForm,
 } from "@redux/slices/scheduleSlice.tsx";
-import { RequestSchedule, Schedule } from "@app/types/schedule.ts";
+import { RequestSchedule } from "@app/types/schedule.ts";
 import { useAppDispatch } from "@redux/hooks.ts";
-import moment from "moment/moment";
 import { v4 as uuidv4 } from "uuid";
 import { useCreateSchedule } from "@app/tanstack-query/schedules/useCreateSchedule.ts";
 import { useUser } from "@app/tanstack-query/useUser.ts";
@@ -32,33 +31,6 @@ const useSchedule = () => {
   });
 
   const schedules = data?.data;
-
-  const todaySchedules =
-    schedules?.filter(
-      (schedule) =>
-        moment(date).isSameOrAfter(schedule.start_date) &&
-        moment(date).isSameOrBefore(schedule.end_date)
-    ) ?? [];
-
-  const lastDay = moment(date).endOf("month").date();
-  const arr = Array.from({ length: lastDay }, (_, i) => i + 1);
-  const init: { [key: string]: Schedule[] } = {};
-  const monthSchedules = arr.reduce((prev, current) => {
-    const date = moment(`${month}-${current}`, "YYYY-MM-D");
-    const today = schedules?.filter(
-      (schedule: Schedule) =>
-        date.isSameOrAfter(schedule.start_date) &&
-        date.isSameOrBefore(schedule.end_date)
-    );
-    if (today && today.length !== 0) {
-      return {
-        ...prev,
-        [date.format("YYYY-MM-DD")]: today,
-      };
-    } else {
-      return prev;
-    }
-  }, init);
 
   const handleCreateSchedule = async (schedule: RequestSchedule) => {
     if (!user) {
@@ -97,10 +69,8 @@ const useSchedule = () => {
   return {
     data,
     schedules,
-    monthSchedules,
     isPending,
     isError,
-    todaySchedules,
     date,
     month,
     handleDeleteSchedule,
